LPS-84125 Copy code to edit structure display context
